refactor: 트림 선택 페이지 컴포넌트들 Components/Custom 에 추가
