+++ front/.vite/deps_temp_2e810e58/package.json
@@ -1,0 +1,3 @@
+{
+  "type": "module"
+}
